vis-inject-v1-docs-and-report/scripts/requirements.txt
python-pptx>=1.0
matplotlib>=3.8
Pillow>=10
